backend/pyproject.toml
[project]
name = "todo-project-name"
version = "0.1.0"
description = "Backend web app"
authors = [
    {name = "Kye Burchard", email = "kyeb@mit.edu"},
]
dependencies = [
    "strawberry-graphql[fastapi]>=0.227.2",
    "fastapi>=0.110.2",
    "uvicorn[standard]>=0.29.0",
    "boto3>=1.34.93",
    "shortuuid>=1.0.13",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
